refactor(jest-cli): reduce nesting in the package
