+++ jest.config.mjs
@@ -39,6 +39,7 @@ export default {
   testPathIgnorePatterns: [
     '/__arbitraries__/',
     '/__benchmarks__/',
+    '/__fixtures__/',
     '/__typetests__/',
     '/node_modules/',
     '/examples/',
@@ -53,20 +54,16 @@ export default {
     '/packages/.*/build',
     '/packages/.*/src/__tests__/setPrettyPrint.ts',
     '/packages/jest-core/src/__tests__/test_root',
-    '/packages/jest-core/src/__tests__/__fixtures__/',
-    '/packages/jest-cli/src/init/__tests__/fixtures/',
     '/packages/jest-haste-map/src/__tests__/haste_impl.js',
     '/packages/jest-haste-map/src/__tests__/dependencyExtractor.js',
     '/packages/jest-haste-map/src/__tests__/test_dotfiles_root/',
     '/packages/jest-repl/src/__tests__/test_root',
-    '/packages/jest-resolve-dependencies/src/__tests__/__fixtures__/',
     '/packages/jest-runtime/src/__tests__/defaultResolver.js',
     '/packages/jest-runtime/src/__tests__/module_dir/',
     '/packages/jest-runtime/src/__tests__/NODE_PATH_dir',
     '/packages/jest-snapshot/src/__tests__/plugins',
     '/packages/jest-snapshot/src/__tests__/fixtures/',
     '/packages/jest-validate/src/__tests__/fixtures/',
-    '/packages/jest-worker/src/workers/__tests__/__fixtures__/',
     '/e2e/__tests__/iterator-to-null-test.ts',
     '/e2e/__tests__/tsIntegration.test.ts', // this test needs types to be build, it runs in a separate CI job through `jest.config.ts.mjs`
   ],
