+++ packages/jest-cli/src/__tests__/args.test.ts
@@ -8,8 +8,8 @@
 
 import type {Config} from '@jest/types';
 import {constants} from 'jest-config';
-import {buildArgv} from '../../cli';
-import {check} from '../../cli/args';
+import {check} from '../args';
+import {buildArgv} from '../run';
 
 const argv = (input: Partial<Config.Argv>): Config.Argv => input as Config.Argv;
