+++ packages/jest-cli/src/init/__tests__/__fixtures__/has-jest-config-file-mjs/package.json
