+++ packages/jest-cli/src/init/__tests__/__fixtures__/has-jest-config-file-ts/jest.config.ts
