+++ packages/jest-cli/src/init/__tests__/__fixtures__/only-package-json/package.json
