+++ packages/jest-cli/src/init/__tests__/__fixtures__/test-generated-jest-config-ts/package.json
