+++ packages/jest-cli/src/init/__tests__/__fixtures__/test-script-configured/package.json
