+++ packages/jest-cli/src/init/__tests__/__fixtures__/type-module/package.json
