+++ packages/jest-cli/src/init/__tests__/__fixtures__/typescript-in-dev-dependencies/package.json
